+++ eslint/config/rulesets/eslint-comments.js
@@ -1,4 +1,10 @@
 /** @type {import('eslint').ESLint.ConfigData} */
 module.exports = {
-  extends: ['plugin:eslint-comments/recommended'],
+  extends: ['plugin:@eslint-community/eslint-comments/recommended'],
+  rules: {
+    '@eslint-community/eslint-comments/no-restricted-disable': [
+      'error',
+      'no-unlimited-disable', // Disallow eslint-disable comments without rule names
+    ],
+  },
 };