+++ prettier.config.js
@@ -1,4 +1,2 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const config = require('./prettier/config/index.js');
-
-module.exports = config;
+module.exports = require('./prettier/config');